+++ samples/Dom.Mediator.Samples.MinimalApi/Features/UpdateTask/UpdateTaskCommand.cs
@@ -2,4 +2,6 @@ using Dom.Mediator.Abstractions;
 
 public record UpdateTaskCommand(
     string Id,
-    Status Status) : ICommand;
+    string Comment,
+    Status? Status
+    ) : ICommand;
