+++ samples/Dom.Mediator.Samples.MinimalApi/Infrastructure/Repositories/TaskStore.cs
@@ -14,9 +14,17 @@ public class TaskItem
     public DateTime? DueDate { get; set; }
     public Status Status { get; set; }
     public DateTime CreatedAt { get; set; }
+    public List<Comment> Comments { get; } = new();
+
+    public void AddComment(string text)
+    {
+        Comments.Add(new Comment(text, DateTime.UtcNow));
+    }
 }
 
 public class TaskStore
 {
     public List<TaskItem> Tasks { get; } = new();
 }
+
+public record Comment(string Text, DateTime CreatedAt);
